check Remarques

+++ src/main/java/com/akinobank/app/repositories/AgenceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AgenceRepository extends JpaRepository<Agence, Integer> {
+public interface AgenceRepository extends JpaRepository<Agence, Long> {
 }
